refactor: address PR review feedback
- Add .js extensions to all import statements for proper ES module compatibility
- Optimize root node lookup using direct Map.get() instead of Array.find() for O(1) performance
- Move helper functions canToggleNode and getCurrentCursorNode to utility module for better organization
- Rename test file from .spec.ts to .test.ts to follow project convention
- Remove unused imports to fix TypeScript compilation warnings
- Apply automatic code formatting

All tests continue to pass ✅
TypeScript compilation successful ✅
Code quality improvements based on AI review feedback
