+++ src/types/collapsible.ts
@@ -2,7 +2,7 @@
  * Types for collapsible JSON viewer functionality
  */
 
-import type { JsonValue } from "./index";
+import type { JsonValue } from "./index.js";
 
 export interface JsonPath {
   type: "root" | "object" | "array";
